chore: User api, domain 모듈 초기 디렉터리 구조 설정
