+++ aics-apis/aics-api/src/main/java/kgu/developers/apis/api/user/application/UserService.java
@@ -1,0 +1,12 @@
+package kgu.developers.apis.api.user.application;
+
+import org.springframework.stereotype.Service;
+
+import kgu.developers.core.domain.user.domain.UserRepository;
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+	private final UserRepository userRepository;
+}
