+++ aics-apis/aics-api/src/main/java/kgu/developers/apis/api/user/presentation/UserController.java
@@ -1,0 +1,7 @@
+package kgu.developers.apis.api.user.presentation;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+}
