+++ aics-apis/aics-api/src/main/java/kgu/developers/apis/api/user/presentation/exception/UserExceptionCode.java
@@ -1,0 +1,19 @@
+package kgu.developers.apis.api.user.presentation.exception;
+
+import kgu.developers.core.common.exception.ExceptionCode;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public enum UserExceptionCode implements ExceptionCode {
+	;
+	private final HttpStatus status;
+	private final String message;
+
+	@Override
+	public String getCode() {
+		return this.name();
+	}
+}
