+++ aics-apis/aics-api/src/main/java/kgu/developers/apis/api/user/presentation/request/UserRequest.java
@@ -1,0 +1,4 @@
+package kgu.developers.apis.api.user.presentation.request;
+
+public class UserRequest {
+}
