+++ aics-core/aics-domain/src/main/java/kgu/developers/core/domain/post/Post.java
@@ -1,20 +1,19 @@
 package kgu.developers.core.domain.post;
 
+import static jakarta.persistence.EnumType.*;
+import static jakarta.persistence.FetchType.*;
+import static jakarta.persistence.GenerationType.*;
+import static lombok.AccessLevel.*;
 
-import static jakarta.persistence.EnumType.STRING;
-import static lombok.AccessLevel.PROTECTED;
-import static jakarta.persistence.FetchType.LAZY;
-import static jakarta.persistence.GenerationType.IDENTITY;
-
-import jakarta.persistence.Id;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Enumerated;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import kgu.developers.core.common.domain.BaseTimeEntity;
-import kgu.developers.core.domain.user.User;
+import kgu.developers.core.domain.user.domain.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -26,27 +25,27 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = PROTECTED)
 public class Post extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    @Column(name = "post_id")
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = IDENTITY)
+	@Column(name = "post_id")
+	private Long id;
 
-    @Column(nullable = false, length = 100)
-    private String title;
+	@Column(nullable = false, length = 100)
+	private String title;
 
-    @Column(nullable = false, length = 3000)
-    private String content;
+	@Column(nullable = false, length = 3000)
+	private String content;
 
-    @Column(nullable = false)
-    private int views;
+	@Column(nullable = false)
+	private int views;
 
-    @Column(nullable = false)
-    @Enumerated(STRING)
-    private Category category;
+	@Column(nullable = false)
+	@Enumerated(STRING)
+	private Category category;
 
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "user_id")
-    private User author;
+	@ManyToOne(fetch = LAZY)
+	@JoinColumn(name = "user_id")
+	private User author;
 
 
     /*
@@ -56,13 +55,7 @@ public class Post extends BaseTimeEntity {
     private File fileID;
     */
 
-    public static Post create(String title, String content, Category category, User author) {
-        return Post.builder()
-                .title(title)
-                .content(content)
-                .views(0)
-                .category(category)
-                .author(author)
-                .build();
-    }
+	public static Post create(String title, String content, Category category, User author) {
+		return Post.builder().title(title).content(content).views(0).category(category).author(author).build();
+	}
 }
