--- aics-core/aics-domain/src/main/java/kgu/developers/core/domain/user/Gender.java
@@ -1,15 +1,0 @@
-package kgu.developers.core.domain.user;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Gender {
-
-    MEN("남자"),
-    WOMEN("여자")
-    ;
-
-    private final String description;
-}
