--- aics-core/aics-domain/src/main/java/kgu/developers/core/domain/user/Grade.java
@@ -1,19 +1,0 @@
-package kgu.developers.core.domain.user;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Grade {
-
-    FRESH("1학년"),
-    SOPHOMORE("2학년"),
-    JUNIOR("3학년"),
-    SENIOR("4학년"),
-    FIFTH("5학년")
-    ;
-
-
-    private final String description;
-}
