+++ aics-core/aics-domain/src/main/java/kgu/developers/core/domain/user/domain/Status.java
@@ -1,0 +1,13 @@
+package kgu.developers.core.domain.user.domain;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Status {
+
+	INSCHOOL("재학중"), EXPELLED("퇴학"), LEAVED("휴학"), ETC("기타");
+
+	private final String description;
+}
