+++ aics-core/aics-domain/src/main/java/kgu/developers/core/domain/user/domain/UserRepository.java
@@ -1,0 +1,4 @@
+package kgu.developers.core.domain.user.domain;
+
+public interface UserRepository {
+}
